added a middleware & edited the menu bar

+++ routes/web.php
@@ -19,11 +19,11 @@ Route::get('/', [ItemController::class ,'index']);
 
 //create items 
 
-Route::get('/item/create',[ItemController::class,'create']);
+Route::get('/item/create',[ItemController::class,'create'])->middleware('auth');
 
 //store items
 
-Route::post('/item',[ItemController::class,'store']);
+Route::post('/item',[ItemController::class,'store'])->middleware('auth');
 
 //view item
 
@@ -31,18 +31,18 @@ Route::get('/item/{item}',[ItemController::class,'show']);
 
 //edit item
 
-Route::get('/item/{item}/edit',[ItemController::class,'edit']);
+Route::get('/item/{item}/edit',[ItemController::class,'edit'])->middleware('auth');
 
 //update item
 
-Route::put('/item/{item}',[ItemController::class,'update']);
+Route::put('/item/{item}',[ItemController::class,'update'])->middleware('auth');
 
 //delete item
-Route::delete('/item/{item}',[ItemController::class,'destroy']);
+Route::delete('/item/{item}',[ItemController::class,'destroy'])->middleware('auth');
 
 
 //register for user
-Route::get('/register',[UserController::class,'create']);
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
 //to store a user
 
@@ -51,13 +51,13 @@ Route::post("/users",[UserController::class,'store']);
 
 //to log out a user
 
-Route::post("/logout",[UserController::class,'logout']);
+Route::post("/logout",[UserController::class,'logout'])->middleware('auth');
 
 //to login a user
 
-Route::get("/login",[UserController::class,'login']);
+Route::get("/login",[UserController::class,'login'])->middleware('guest')->name('login');
 
 
 ///user authentcation
 
-Route::post("/user/authentcation",[UserController::class,'authentcation']);
+Route::post("/user/authentcation",[UserController::class,'authentcation'])->middleware('guest');
